fix(core): split the ISO 8859 code-page tables out of step_encoding.rs
The code-page tables added for the S-escape page-selection fix tipped
step_encoding.rs to 426 lines, over the repo's 400-line module-size
ratchet (rust/processing/tests/module_size_ratchet.rs), which failed
PR CI ("New non-generated .rs file(s) over 400 lines with no allowlist
row").

The tables are inert data with no behavioural seam of their own, so
split by extraction rather than allowlisting: moved CODEPAGE_TABLES and
resolve_extended_char into a new sibling module, step_codepages.rs
(pub(crate), no public surface), wired in via a path-attributed mod
declaration in step_encoding.rs.

step_encoding.rs: 326 lines, step_codepages.rs: 111 lines - both well
under budget. No behaviour change; the TS side (ifc-string.ts, 323
lines) was already under budget and needed no change.
